+++ examples/2.3.2-mlt-stream.php
@@ -3,6 +3,8 @@
 require(__DIR__.'/init.php');
 htmlHeader();
 
+echo "<h2>Note: The techproducts example doesn't include a /mlt handler anymore!</h2>";
+
 // create a client instance
 $client = new Solarium\Client($config);